Entrega3_Punto4

+++ SII_App_Grupo_5/SII_App_Grupo_5/Models/Enajenante.cs
@@ -10,7 +10,7 @@ namespace SII_App_Grupo_5.Models
         public int InscripcionId { get; set; }
         virtual public Inscripcion Inscripcion { get; set; }
         public string Rut { get; set; }
-        public int PorcentajeDerecho { get; set; }
+        public float PorcentajeDerecho { get; set; }
         public bool Acreditado { get; set; }
     }
 }
